+++ cpp/src/experimental/sssp.cu
@@ -42,7 +42,7 @@ namespace experimental {
 namespace detail {
 
 template <typename GraphViewType, typename PredecessorIterator>
-void sssp(raft::handle_t &handle,
+void sssp(raft::handle_t const &handle,
           GraphViewType const &push_graph_view,
           typename GraphViewType::weight_type *distances,
           PredecessorIterator predecessor_first,
@@ -241,7 +241,7 @@ void sssp(raft::handle_t &handle,
 }  // namespace detail
 
 template <typename vertex_t, typename edge_t, typename weight_t, bool multi_gpu>
-void sssp(raft::handle_t &handle,
+void sssp(raft::handle_t const &handle,
           graph_view_t<vertex_t, edge_t, weight_t, false, multi_gpu> const &graph_view,
           weight_t *distances,
           vertex_t *predecessors,
@@ -265,7 +265,7 @@ void sssp(raft::handle_t &handle,
 
 // explicit instantiation
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int32_t, float, false, true> const &graph_view,
                    float *distances,
                    int32_t *predecessors,
@@ -273,7 +273,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int32_t, double, false, true> const &graph_view,
                    double *distances,
                    int32_t *predecessors,
@@ -281,7 +281,7 @@ template void sssp(raft::handle_t &handle,
                    double cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int64_t, float, false, true> const &graph_view,
                    float *distances,
                    int32_t *predecessors,
@@ -289,7 +289,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int64_t, double, false, true> const &graph_view,
                    double *distances,
                    int32_t *predecessors,
@@ -297,7 +297,7 @@ template void sssp(raft::handle_t &handle,
                    double cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int64_t, int64_t, float, false, true> const &graph_view,
                    float *distances,
                    int64_t *predecessors,
@@ -305,7 +305,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int64_t, int64_t, double, false, true> const &graph_view,
                    double *distances,
                    int64_t *predecessors,
@@ -313,7 +313,7 @@ template void sssp(raft::handle_t &handle,
                    double cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int32_t, float, false, false> const &graph_view,
                    float *distances,
                    int32_t *predecessors,
@@ -321,7 +321,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int32_t, double, false, false> const &graph_view,
                    double *distances,
                    int32_t *predecessors,
@@ -329,7 +329,7 @@ template void sssp(raft::handle_t &handle,
                    double cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int64_t, float, false, false> const &graph_view,
                    float *distances,
                    int32_t *predecessors,
@@ -337,7 +337,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int32_t, int64_t, double, false, false> const &graph_view,
                    double *distances,
                    int32_t *predecessors,
@@ -345,7 +345,7 @@ template void sssp(raft::handle_t &handle,
                    double cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int64_t, int64_t, float, false, false> const &graph_view,
                    float *distances,
                    int64_t *predecessors,
@@ -353,7 +353,7 @@ template void sssp(raft::handle_t &handle,
                    float cutoff,
                    bool do_expensive_check);
 
-template void sssp(raft::handle_t &handle,
+template void sssp(raft::handle_t const &handle,
                    graph_view_t<int64_t, int64_t, double, false, false> const &graph_view,
                    double *distances,
                    int64_t *predecessors,