+++ python/cugraph/dask/traversal/mg_sssp_wrapper.pyx
@@ -15,26 +15,25 @@
 #
 
 from cugraph.structure.utils_wrapper import *
-from cugraph.dask.traversal cimport mg_bfs as c_bfs
+from cugraph.dask.traversal cimport mg_sssp as c_sssp
 import cudf
 from cugraph.structure.graph_primtypes cimport *
 import cugraph.structure.graph_primtypes_wrapper as graph_primtypes_wrapper
 from libc.stdint cimport uintptr_t
 
-def mg_bfs(input_df,
-           num_global_verts,
-           num_global_edges,
-           vertex_partition_offsets,
-           rank,
-           handle,
-           start,
-           return_distances=False):
+def mg_sssp(input_df,
+            num_global_verts,
+            num_global_edges,
+            vertex_partition_offsets,
+            rank,
+            handle,
+            start):
     """
-    Call pagerank
+    Call sssp
     """
 
     cdef size_t handle_size_t = <size_t>handle.getHandle()
-    handle_ = <c_bfs.handle_t*>handle_size_t
+    handle_ = <c_sssp.handle_t*>handle_size_t
 
     # Local COO information
     src = input_df['src']
@@ -48,6 +47,7 @@ def mg_bfs(input_df,
         weights = input_df['value']
         weight_t = weights.dtype
     else:
+        weights = None
         weight_t = np.dtype("float32")
 
     # FIXME: Offsets and indices are currently hardcoded to int, but this may
@@ -63,6 +63,8 @@ def mg_bfs(input_df,
     cdef uintptr_t c_src_vertices = src.__cuda_array_interface__['data'][0]
     cdef uintptr_t c_dst_vertices = dst.__cuda_array_interface__['data'][0]
     cdef uintptr_t c_edge_weights = <uintptr_t>NULL
+    if weights is not None:
+        c_edge_weights = weights.__cuda_array_interface__['data'][0]
 
     # FIXME: data is on device, move to host (to_pandas()), convert to np array and access pointer to pass to C
     vertex_partition_offsets_host = vertex_partition_offsets.values_host
@@ -85,24 +87,29 @@ def mg_bfs(input_df,
     # Generate the cudf.DataFrame result
     df = cudf.DataFrame()
     df['vertex'] = cudf.Series(np.arange(vertex_partition_offsets.iloc[rank], vertex_partition_offsets.iloc[rank+1]), dtype=vertex_t)
-    df['predecessor'] = cudf.Series(np.zeros(len(df['vertex']), dtype=np.int32))
-    if (return_distances):
-        df['distance'] = cudf.Series(np.zeros(len(df['vertex']), dtype=np.int32))
+    df['predecessor'] = cudf.Series(np.zeros(len(df['vertex']), dtype=vertex_t))
+    df['distance'] = cudf.Series(np.zeros(len(df['vertex']), dtype=weight_t))
 
     # Associate <uintptr_t> to cudf Series
-    cdef uintptr_t c_distance_ptr    = <uintptr_t> NULL # Pointer to the DataFrame 'distance' Series
     cdef uintptr_t c_predecessor_ptr = df['predecessor'].__cuda_array_interface__['data'][0]
-    if (return_distances):
-        c_distance_ptr = df['distance'].__cuda_array_interface__['data'][0]
+    cdef uintptr_t c_distance_ptr = df['distance'].__cuda_array_interface__['data'][0]
 
-    cdef bool direction = <bool> 1
     # MG BFS path assumes directed is true
-    c_bfs.call_bfs[int, float](handle_[0],
-                               graph_container,
-                               <int*> NULL,
-                               <int*> c_distance_ptr,
-                               <int*> c_predecessor_ptr,
-                               <double*> NULL,
-                               <int> start,
-                               direction)
+    if weight_t == np.float32:
+        c_sssp.call_sssp[int, float](handle_[0],
+                                     graph_container,
+                                     <int*> NULL,
+                                     <float*> c_distance_ptr,
+                                     <int*> c_predecessor_ptr,
+                                     <int> start)
+    elif weight_t == np.float64:
+        c_sssp.call_sssp[int, double](handle_[0],
+                                      graph_container,
+                                      <int*> NULL,
+                                      <double*> c_distance_ptr,
+                                      <int*> c_predecessor_ptr,
+                                      <int> start)
+    else: # This case should not happen
+        raise NotImplementedError
+
     return df