+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - PR #1163 Integrated 2D shuffling and Louvain updates
 - PR #1178 Refactored cython graph factory code to scale to additional data types
 - PR #1175 Integrated 2D pagerank python/cython infra
+- PR #1177 Integrated 2D bfs and sssp python/cython infra
 
 ## Improvements
 - PR 1081 MNMG Renumbering - sort partitions by degree
@@ -48,6 +49,7 @@
 - PR #1192 Fix benchmark notes and documentation issues in graph.py
 - PR #1196 Move subcomms init outside of individual algorithm functions
 - PR #1198 Remove deprecated call to from_gpu_matrix
+
 
 # cuGraph 0.15.0 (26 Aug 2020)